fix: split gc into separate file to make testing easier
