+++ index.js
@@ -1,13 +1,8 @@
 'use strict'
 
 const EventEmitter = require('events')
-const { monitorEventLoopDelay, PerformanceObserver, constants } = require('perf_hooks')
-const {
-  NODE_PERFORMANCE_GC_MAJOR,
-  NODE_PERFORMANCE_GC_MINOR,
-  NODE_PERFORMANCE_GC_INCREMENTAL,
-  NODE_PERFORMANCE_GC_WEAKCB
-} = constants
+const { monitorEventLoopDelay } = require('perf_hooks')
+const GCStats = require('./gc')
 
 function hrtime2ns (time) {
   return time[0] * 1e9 + time[1]
@@ -36,59 +31,6 @@ function validate ({
   }
 }
 
-class GCStats {
-  static entryKindToKey (entry) {
-    switch (entry.kind) {
-      case NODE_PERFORMANCE_GC_MAJOR:
-        return 'major'
-      case NODE_PERFORMANCE_GC_MINOR:
-        return 'minor'
-      case NODE_PERFORMANCE_GC_INCREMENTAL:
-        return 'incremental'
-      case NODE_PERFORMANCE_GC_WEAKCB:
-        return 'weakCB'
-      default:
-        return null
-    }
-  }
-
-  constructor () {
-    this.reset()
-  }
-
-  reset () {
-    // These are set as a tuple of [count; rollingAverage]
-    this.major = [0, undefined]
-    this.minor = [0, undefined]
-    this.incremental = [0, undefined]
-    this.weakCB = [0, undefined]
-  }
-
-  update (gcEntry) {
-    const key = GCStats.entryKindToKey(gcEntry)
-    if (key) {
-      let [count, average] = this[key]
-      if (average === undefined) {
-        this[key] = [1, gcEntry.duration]
-      } else {
-        count++
-        average -= average / count
-        average += gcEntry.duration / count
-        this[key] = [count, average]
-      }
-    }
-  }
-
-  data () {
-    return {
-      major: this.major[1],
-      minor: this.minor[1],
-      incremental: this.incremental[1],
-      weakCB: this.weakCB[1]
-    }
-  }
-}
-
 class Doc extends EventEmitter {
   constructor (options) {
     super()
@@ -108,15 +50,7 @@ class Doc extends EventEmitter {
       memory: {}
     }
 
-    const gcStats = new GCStats()
-    const gcObserver = new PerformanceObserver(list => {
-      for (const gcEntry of list.getEntriesByType('gc')) {
-        gcStats.update(gcEntry)
-      }
-    })
-    gcObserver.observe({
-      entryTypes: ['gc']
-    })
+    const gcStats = GCStats()
 
     let lastCheck = process.hrtime()
     let cpuUsage = process.cpuUsage()
